+++ Galapagos/galapagos.h
@@ -4,10 +4,8 @@
 #include <cstdint>
 
 #include "stochastic.h"
-#include "galapagos_types.h"
+#include "galapagos_metadata.h"
 #include "population.h"
-#include "population_metadata.h"
-#include "creature.h"
 
 class galapagos_session {
 public:
@@ -25,11 +23,11 @@ public:
 
 GALAPAGOS_API population_metadata* create_population_metadata();
 
-GALAPAGOS_API void set_population_metadata_size(population_metadata* population_metadata, uint32_t size);
-GALAPAGOS_API uint32_t get_population_metadata_size(population_metadata* population_metadata);
+GALAPAGOS_API void set_population_metadata_size(population_metadata* population_metadata, size_t size);
+GALAPAGOS_API size_t get_population_metadata_size(population_metadata* population_metadata);
 
-GALAPAGOS_API void set_population_metadata_survival_rate(population_metadata* population_metadata, uint32_t survival_rate);
-GALAPAGOS_API uint32_t get_population_metadata_survival_rate(population_metadata* population_metadata);
+GALAPAGOS_API void set_population_metadata_survival_rate(population_metadata* population_metadata, double survival_rate);
+GALAPAGOS_API double get_population_metadata_survival_rate(population_metadata* population_metadata);
 
 GALAPAGOS_API population* create_population(population_metadata* population_metadata);
